fix: only set metadata once
