+++ src/data-workspace/data-workspace.js
@@ -41,6 +41,30 @@ const dataValueQuery = {
     },
 }
 
+const metadataQuery = {
+    metadata: {
+        resource: 'metadata',
+        params: {
+            // Note: on dataSet.dataSetElement, the categoryCombo property is
+            // included because it can mean it's overriding the data element's
+            // native categoryCombo. It can sometimes be absent from the data
+            // set element
+            'dataSets:fields':
+                'id,displayFormName,formType,dataSetElements[dataElement,categoryCombo],categoryCombo,sections~pluck',
+            'dataElements:fields': 'id,displayFormName,categoryCombo,valueType',
+            'sections:fields':
+                'id,displayName,sortOrder,showRowTotals,showColumnTotals,disableDataElementAutoGroup,greyedFields[id],categoryCombos~pluck,dataElements~pluck,indicators~pluck',
+            'categoryCombos:fields':
+                'id,skipTotal,categories~pluck,categoryOptionCombos~pluck,isDefault',
+            'categories:fields': 'id,displayFormName,categoryOptions~pluck',
+            'categoryOptions:fields':
+                'id,displayFormName,categoryOptionCombos~pluck,categoryOptionGroups~pluck,isDefault',
+            'categoryOptionCombos:fields':
+                'id,categoryOptions~pluck,categoryCombo,name',
+        },
+    },
+}
+
 const useDataValues = () => {
     const [{ dataSetId, orgUnitId, periodId }] = useContextSelection()
     const attributeOptionComboId = useAttributeOptionCombo()
@@ -77,7 +101,6 @@ function mapDataValuesToFormInitialValues(dataValues) {
 const useAttributeOptionCombo = () => {
     const { available, metadata } = useMetadata()
     const [{ dataSetId, attributeOptionComboSelection }] = useContextSelection()
-
     const cocId = useMemo(() => {
         if (available && dataSetId) {
             const dataSet = getDataSetById(metadata, dataSetId)
@@ -117,7 +140,12 @@ export const DataWorkspace = () => {
         },
     ])
 
-    const { available, metadata } = useMetadata()
+    const { available, metadata, setMetadata } = useMetadata()
+    useQuery([metadataQuery], {
+        staleTime: 60 * 24 * 1000,
+        refetchOnWindowFocus: false,
+        onSuccess: (data) => setMetadata(data.metadata),
+    })
 
     const dataValuesFetch = useDataValues()
     console.log({ metadata }, { dataValues: dataValuesFetch.data }, { dataSet: dataSetFetch.data })
