+++ src/data-workspace/metadata-context.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import React, { useState, useMemo, useContext } from 'react'
+import React, { useState, useMemo, useContext, useEffect } from 'react'
 import { useQuery } from 'react-query'
 import { hashArraysInObject } from './utils.js'
 
@@ -24,7 +24,9 @@ export const MetadataContextProvider = ({ children }) => {
     const [loading, setLoading] = useState(true)
 
     const setMetadata = (value) => {
-        setContext(value)
+        const hashed = hashArraysInObject(value)
+        console.log('set metadata', hashed)
+        setContext(hashed)
         setLoading(false)
         localStorage.setItem('metadata', JSON.stringify(value))
     }
@@ -51,42 +53,4 @@ MetadataContextProvider.propTypes = {
     children: PropTypes.node,
 }
 
-const metadataQuery = {
-    metadata: {
-        resource: 'metadata',
-        params: {
-            // Note: on dataSet.dataSetElement, the categoryCombo property is
-            // included because it can mean it's overriding the data element's
-            // native categoryCombo. It can sometimes be absent from the data
-            // set element
-            'dataSets:fields':
-                'id,displayFormName,formType,dataSetElements[dataElement,categoryCombo],categoryCombo,sections~pluck',
-            'dataElements:fields': 'id,displayFormName,categoryCombo,valueType',
-            'sections:fields':
-                'id,displayName,sortOrder,showRowTotals,showColumnTotals,disableDataElementAutoGroup,greyedFields[id],categoryCombos~pluck,dataElements~pluck,indicators~pluck',
-            'categoryCombos:fields':
-                'id,skipTotal,categories~pluck,categoryOptionCombos~pluck,isDefault',
-            'categories:fields': 'id,displayFormName,categoryOptions~pluck',
-            'categoryOptions:fields':
-                'id,displayFormName,categoryOptionCombos~pluck,categoryOptionGroups~pluck,isDefault',
-            'categoryOptionCombos:fields':
-                'id,categoryOptions~pluck,categoryCombo,name',
-        },
-    },
-}
-
-export const useMetadata = () => {
-    const context = useContext(MetadataContext)
-
-    useQuery([metadataQuery], {
-        onSuccess: (metadata) => {
-            const hashed =
-                metadata?.metadata && hashArraysInObject(metadata.metadata)
-            context.setMetadata(hashed)
-        },
-        staleTime: 60 * 24 * 1000,
-        refetchOnWindowFocus: false,
-    })
-
-    return context
-}
+export const useMetadata = () => useContext(MetadataContext)
